Refactor Parsers into more descriptive packages
